% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReweightDB.R
\name{computeTable1LikeTransformation}
\alias{computeTable1LikeTransformation}
\title{Compute Table 1 like transformation}
\usage{
computeTable1LikeTransformation(X, outcomeBalance, outcomeCol = "Y")
}
\arguments{
\item{X}{data frame}

\item{outcomeBalance}{boolean specifying if to use outcome in balancing}

\item{outcomeCol}{string specifying the name of the outcome column}
}
\value{
A data frame
}
\description{
Compute features and squared numeric feature multiplied by outcome and 1-outcome.
}
