% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReweightDB.R
\name{dualReweightByMeans}
\alias{dualReweightByMeans}
\title{Reweight an internal database to match the means of an external one using
a dual formulation of a weight optimization problem.}
\usage{
dualReweightByMeans(X, mu, lambda, minSd, minW, verbose)
}
\arguments{
\item{mu}{a vector of means of the internal dataset.}

\item{lambda}{lambda - regularization parameter}

\item{minSd}{minimum variance for a columns to be included}

\item{Z}{a data frame where every row stores a sample of the internal
databse.}
}
\value{
A vector of weights
}
\description{
This function recieves a data frame Z of an internal DB and a vector mu of
means from an external DBs. The elements of mu correspond to columns of Z.
It returns a set of sample weights such that the weighted means of the
columns of Z are as close as possible to the elements of mu while
minimizing the divergence between the distribution of the weights and the
uniform distribution.
}
