% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReweightDB.R
\name{normalizeDataAndExpectations}
\alias{normalizeDataAndExpectations}
\title{Normalize data and expectations}
\usage{
normalizeDataAndExpectations(Z, mu, minSd)
}
\arguments{
\item{Z}{data frame}

\item{mu}{vector of expectations}

\item{minSd}{float value of minimum standard deviation. Columns with a smaller sd will be removed}
}
\value{
list
}
\description{
Normalize data and expectations.
}
