% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReweightDB.R
\name{primalReweightByMeans}
\alias{primalReweightByMeans}
\title{Reweight an internal database to match the means of an external via solving a corresponding optimization problem.}
\usage{
primalReweightByMeans(
  Z,
  mu,
  divergence,
  lambda,
  minSd,
  minW,
  distance,
  verbose
)
}
\value{
A vector of weights
}
\description{
This function recieves a data frame Z of an internal DB and a vector mu of
means from an external DBs. The elements of mu correspond to columns of Z.
It returns a set of sample weights such that the weighted means of the
columns of Z are as close as possible to the elements of mu while
minimizing the divergence between the distribution of the weights and the
uniform distribution.

SEE THE PARENT FUNCTION
}
