% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReweightDB.R
\docType{package}
\name{ReweightDB}
\alias{ReweightDB}
\title{Reweight an internal database to match the means of an external one.}
\description{
A package for reweighting a dataset to match the statistics of an external one
}
\details{
ReweightDB
}
